backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
requests==2.32.3
python-dotenv==1.0.1
werkzeug==3.0.3
pdfplumber
# psycopg2-binary  ← للـ production فقط (PostgreSQL)
# محلياً نستخدم SQLite (لا يحتاج driver خارجي)
